-R/--syncnt: set number of SYN retransmits
Add an option to set the number of SYN retransmits for connect(2):

        # 1 + 2 SYN retransmits
        GLB_OPTIONS="--syncnt 2 --single ..."

libglb.so uses a blocking connect(2) with the default timeouts. On
Linux, libglb will send 7 SYN requests, taking around 2 minutes to
failover.

If the process uses non-blocking sockets, libglb converts the socket
to non-blocking mode before calling connect(2) and resets the socket
after the call returns. The intercepted process will not be able to
perform other operations until connect(2) returns.

With services on a local network, connect(2) usually either quickly
succeeds or is reset. If the operation times out, the intercepted
service will appear to be hung for up to 2 minutes.

The options for setting a timeout on a blocking connect(2) are Linux
specific:

* SO_SNDTIMEO socket option

~~~
struct timeval timeout = {0};
timeout.tv_sec = 10;
setsockopt(sockfd, SOL_SOCKET, SO_SNDTIMEO, &timeout, sizeof(timeout));
~~~

  The behaviour is undefined for connect(2) on most OS'es. Linux treats
  it similarly to a non-blocking socket: when the timeout is reached,
  connect(2) returns with errno set to EINPROGRESS. The in-flight
  connect is not interrupted.

* TCP_SYNCNT socket option

  Another Linux specific option allows setting the number of SYN
  retransmits for the socket. By default, 7 SYN requests are sent taking
  about 2 minutes for the connect to timeout.

  SYN requests are sent with a backoff:

~~~
0 seconds: SYN 1
1 second: SYN 2
3 seconds: SYN 3
7 seconds: SYN 4
15 seconds: SYN 5
31 seconds: SYN 6
63 seconds: SYN 7
~~~